restructure rule on __init__.py && reorganize docs
