+++ chainercv/links/__init__.py
@@ -1,7 +1,1 @@
-from chainercv.links.model.faster_rcnn.faster_rcnn import FasterRCNNBase  # NOQA
 from chainercv.links.model.faster_rcnn.faster_rcnn_vgg import FasterRCNNVGG16  # NOQA
-from chainercv.links.model.faster_rcnn.region_proposal_network import RegionProposalNetwork  # NOQA
-from chainercv.links.model.faster_rcnn.utils.bbox2loc import bbox2loc  # NOQA
-from chainercv.links.model.faster_rcnn.utils.generate_anchor_base import generate_anchor_base  # NOQA
-from chainercv.links.model.faster_rcnn.utils.loc2bbox import loc2bbox  # NOQA
-from chainercv.links.model.faster_rcnn.utils.proposal_creator import ProposalCreator  # NOQA
