+++ tests/links_tests/model_tests/faster_rcnn_tests/test_region_proposal_network.py
@@ -6,7 +6,7 @@ from chainer import cuda
 from chainer import testing
 from chainer.testing import attr
 
-from chainercv.links import RegionProposalNetwork
+from chainercv.links.model.faster_rcnn import RegionProposalNetwork
 
 
 @testing.parameterize(
