+++ tests/links_tests/model_tests/faster_rcnn_tests/utils_tests/test_bbox2loc_loc2bbox.py
@@ -6,8 +6,8 @@ from chainer import cuda
 from chainer import testing
 from chainer.testing import attr
 
-from chainercv.links import bbox2loc
-from chainercv.links import loc2bbox
+from chainercv.links.model.faster_rcnn import bbox2loc
+from chainercv.links.model.faster_rcnn import loc2bbox
 
 
 def generate_bbox(n, img_size, min_length, max_length):
